fix: :hammer: Add src/block/movement.rs
Issue: #21
